***** fdutils version 5.2 diff of 1997/11/04 *****
One of the earliest patched 5.2 version that was
delivered with Debian up to 2.0 (Hamm). This commit
is for historical reasons only. The source code
depends on very first Linux 2.0 kernel sources.

Changes since 5.2:

v19970411
  * Added error check to capacity measurement.
  * Corrected documentation Makefile bug.

v19970629
  * Added vfat support to fdmount.

v19970609
  * Fixed bug where a skew was erroneously applied to a 2m track.
  * Fixed typo in printfdprm.

v19970604
  * Added vfat option for fdmount.

Originated-by: fdutils_5.2pl4.orig.tar.gz
Originated-md5: a391362824a5ad7d4e43433ccec7a0e4
Originated-URL: http://debian.inet.de/debian-archive/debian/dists/hamm/main/source/utils/fdutils_5.2pl4.orig.tar.gz

+++ src/floppycontrol.c
@@ -416,7 +416,7 @@ void main( int argc, char **argv)
 		printf("need_configure=%x\n", fdcstat.need_configure);
 		printf("has_fifo=%x\n", fdcstat.has_fifo);
 		printf("perp_mode=%x\n", fdcstat.perp_mode);
-		printf("address=%x\n", fdcstat.address);
+		printf("address=%x\n", (unsigned int) fdcstat.address);
 	}
 
 #ifdef FDWERRORGET
